+++ test/radar/matchers/class_matcher_test.rb
@@ -6,16 +6,21 @@ class ClassMatcherTest < Test::Unit::TestCase
       @klass = Radar::Matchers::ClassMatcher
     end
 
-    should "match if the class matches" do
+    should "match if the class match exactly" do
+      standard_event = create_exception_event { raise StandardError.new("An error") }
       assert @klass.new(RuntimeError).matches?(create_exception_event)
+      assert !@klass.new(RuntimeError).matches?(standard_event)
     end
 
-    should "not match non-matching classes" do
-      event = create_exception_event do
-        raise StandardError.new("An error")
-      end
+    should "match regular expressions properly" do
+      standard_event = create_exception_event { raise StandardError.new("An error") }
+      assert @klass.new(/.*Error/).matches?(create_exception_event)
+      assert @klass.new(/.*Error/).matches?(standard_event)
+    end
 
-      assert !@klass.new(RuntimeError).matches?(event)
+    should "match subclasses if specified" do
+      assert @klass.new(Exception, :include_subclasses => true).matches?(create_exception_event)
+      assert !@klass.new(Exception).matches?(create_exception_event)
     end
   end
 end