+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.0 (unreleased)
 
+  - Built in matcher: `:class` (or `ClassMatcher`) which checks against the
+    exception class. [GH-17]
   - Add `config.match` to conditionally match exceptions before reporting
     them so that exceptions can be better filtered per application. [GH-11]
   - Changed the way reporters and data extensions are enabled. You must now